♻️ Refactor: Detail 페이지 컴포넌트화 (#5)
