+++ src/styles/detail/DetailPage.module.css
@@ -1,12 +1,17 @@
 .detail-main {
   display: flex;
+  gap: 1.875rem;
 }
 
 .left-section {
   display: flex;
   flex-direction: column;
   flex: 2.5;
-  margin-right: 1.875rem;
+}
+
+.left-section > :nth-child(2) {
+  width: 100%;
+  margin-bottom: 5rem;
 }
 
 .right-section {
@@ -15,33 +20,9 @@
   flex: 1;
 }
 
-.video-section {
-  width: 100%;
-  margin-bottom: 5rem;
-}
-
-.related-title {
-  font-size: 1.5rem;
-  font-weight: bold;
-  margin-bottom: 1.5625rem;
-}
-
-.related-videos {
-  /* overflow-y: scroll; */
-  /* display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(350px, 1fr));
-  margin: 25px auto 50px;
-  gap: 3.125rem 1.25rem; */
-}
-
 @media (max-width: 1050px) {
   .detail-main {
+    gap: 5rem;
     flex-direction: column;
-  }
-  .left-section {
-    margin-right: 0;
-  }
-  .left-section:first-child {
-    margin-bottom: 50px;
   }
 }
