+++ src/styles/detail/RelatedSection.module.css
@@ -1,0 +1,5 @@
+.related-title {
+  font-size: 1.5rem;
+  font-weight: bold;
+  margin-bottom: 1.5625rem;
+}
