+++ src/styles/detail/Video.module.css
@@ -1,6 +1,5 @@
 .video-container {
   /* width: 90%; */
-  min-width: 700px;
   display: flex;
   flex-direction: column;
   margin: auto;
